requirements.txt
json5
requests
colr
websocket-client
rich
delorean